+++ src/main/java/roomescape/LoginMemberArgumentResolver.java
@@ -7,7 +7,6 @@ import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
-import roomescape.auth.AuthorizationExtractor;
 import roomescape.member.Member;
 import roomescape.member.MemberDao;
 import roomescape.member.MemberService;
