+++ src/main/java/roomescape/auth/JwtConfig.java
@@ -1,0 +1,13 @@
+package roomescape.auth;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class JwtConfig {
+
+    @Bean
+    public JwtUtils jwtUtils() {
+        return new JwtUtils();
+    }
+}
