+++ src/main/java/roomescape/auth/TokenProvider.java
@@ -1,23 +1,23 @@
-package roomescape.auth;
-
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.security.Keys;
-import org.springframework.stereotype.Component;
-import roomescape.member.Member;
-
-import java.nio.charset.StandardCharsets;
-
-@Component
-public class TokenProvider {
-    private String secretKey = "Yn2kjibddFAWtnPJ2AFlL8WXmohJMCvigQggaEypa5E=";
-
-    public String createToken(Member member) {
-        String accessToken = Jwts.builder()
-                .setSubject(member.getId().toString())
-                .claim("name", member.getName())
-                .claim("role", member.getRole())
-                .signWith(Keys.hmacShaKeyFor(secretKey.getBytes()))
-                .compact();
-        return accessToken;
-    }
-}
+//package roomescape.auth;
+//
+//import io.jsonwebtoken.Jwts;
+//import io.jsonwebtoken.security.Keys;
+//import org.springframework.stereotype.Component;
+//import roomescape.member.Member;
+//
+//import java.nio.charset.StandardCharsets;
+//
+//@Component
+//public class TokenProvider {
+//    private String secretKey = "Yn2kjibddFAWtnPJ2AFlL8WXmohJMCvigQggaEypa5E=";
+//
+//    public String createToken(Member member) {
+//        String accessToken = Jwts.builder()
+//                .setSubject(member.getId().toString())
+//                .claim("name", member.getName())
+//                .claim("role", member.getRole())
+//                .signWith(Keys.hmacShaKeyFor(secretKey.getBytes()))
+//                .compact();
+//        return accessToken;
+//    }
+//}
