+++ src/main/java/roomescape/member/MemberService.java
@@ -3,18 +3,25 @@ package roomescape.member;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.stereotype.Service;
-import roomescape.auth.AuthorizationExtractor;
+//import roomescape.auth.AuthorizationExtractor;
+import roomescape.auth.JwtUtils;
 
 import java.util.Arrays;
 
 @Service
 public class MemberService {
     private MemberDao memberDao;
-    private AuthorizationExtractor  authorizationExtractor;
+//    private AuthorizationExtractor  authorizationExtractor;
+    private JwtUtils jwtUtils;
 
-    public MemberService(MemberDao memberDao, AuthorizationExtractor authorizationExtractor) {
+//    public MemberService(MemberDao memberDao, AuthorizationExtractor authorizationExtractor) {
+//        this.memberDao = memberDao;
+//        this.authorizationExtractor =  authorizationExtractor;
+//    }
+
+    public MemberService(MemberDao memberDao, JwtUtils jwtUtils) {
         this.memberDao = memberDao;
-        this.authorizationExtractor =  authorizationExtractor;
+        this.jwtUtils = jwtUtils;
     }
 
     public MemberResponse createMember(MemberRequest memberRequest) {
@@ -25,7 +32,7 @@ public class MemberService {
     public Member getMemberFromCookie(HttpServletRequest httpServletRequest) {
         Cookie[] cookies  = httpServletRequest.getCookies();
         String token = Arrays.stream(cookies).filter(cookie -> cookie.getName().equals("token")).findFirst().map(Cookie::getValue).orElseThrow();
-        Member member = memberDao.findById(authorizationExtractor.extractMemberId(token));
+        Member member = memberDao.findById(jwtUtils.extractMemberId(token));
         return member;
     }
 }
