+++ src/main/resources/application.properties
@@ -7,5 +7,6 @@ spring.jpa.show-sql=true
 spring.jpa.properties.hibernate.format_sql=true
 spring.jpa.ddl-auto=create-drop
 spring.jpa.defer-datasource-initialization=true
+spring.main.allow-bean-definition-overriding=true
 
 #roomescape.auth.jwt.secret= Yn2kjibddFAWtnPJ2AFlL8WXmohJMCvigQggaEypa5E=
